# Performance refactor - elide async and await