--- a/src/Microsoft.OData.Core/Json/JsonWriterAsync.cs
+++ b/src/Microsoft.OData.Core/Json/JsonWriterAsync.cs
@@ -332,8 +332,6 @@ namespace Microsoft.OData.Json
                 {
                     return this.writer.WriteAsync(JsonConstants.ArrayElementSeparator);
                 }
-
-                
             }
 
             return TaskUtils.CompletedTask;