… src/Microsoft.OData.Core/Json/ODataJsonReaderCoreUtils.cs
… src/Microsoft.OData.Core/Json/ODataJsonWriterUtils.cs
… src/Microsoft.OData.Core/JsonLight/ODataJsonLightCollectionSerializer.cs
… src/Microsoft.OData.Core/JsonLight/ODataJsonLightEntityReferenceLinkSerializer.cs
… src/Microsoft.OData.Core/JsonLight/ODataJsonLightOutputContext.cs
… src/Microsoft.OData.Core/JsonLight/ODataJsonLightPropertySerializer.cs
… src/Microsoft.OData.Core/JsonLight/ODataJsonLightResourceSerializer.cs
… src/Microsoft.OData.Core/JsonLight/ODataJsonLightSerializer.cs
… src/Microsoft.OData.Core/JsonLight/ODataJsonLightServiceDocumentSerializer.cs
… src/Microsoft.OData.Core/JsonLight/ODataJsonLightValueSerializer.cs
… src/Microsoft.OData.Core/JsonLight/ODataJsonLightWriter.cs
… src/Microsoft.OData.Core/MultipartMixed/ODataMultipartMixedBatchOutputContext.cs
… src/Microsoft.OData.Core/MultipartMixed/ODataMultipartMixedBatchWriter.cs
… src/Microsoft.OData.Core/ODataAsynchronousWriter.cs
… src/Microsoft.OData.Core/ODataBinaryStreamReader.cs
… src/Microsoft.OData.Core/ODataBinaryStreamWriter.cs
… src/Microsoft.OData.Core/ODataCollectionWriterCore.cs
… src/Microsoft.OData.Core/ODataMessage.cs
… src/Microsoft.OData.Core/ODataMessageReader.cs
… src/Microsoft.OData.Core/ODataMessageWriter.cs
… src/Microsoft.OData.Core/ODataMetadataFormat.cs
… src/Microsoft.OData.Core/ODataParameterReaderCoreAsync.cs
… src/Microsoft.OData.Core/ODataParameterWriterCore.cs
… src/Microsoft.OData.Core/ODataReaderCore.cs
… src/Microsoft.OData.Core/ODataUtils.cs
… src/Microsoft.OData.Core/ODataWriter.cs
… src/Microsoft.OData.Core/ODataWriterCore.cs
… src/Microsoft.OData.Core/RawValueWriter.cs
… src/Microsoft.OData.Core/TaskUtils.cs
… test/PublicApiTests/BaseLine/Microsoft.OData.PublicApi.net45.bsl