<!-- markdownlint-disable MD002 MD041 -->

### Description
The following sections describe the nature of changes made to improve performance and memory usage

#### 1. Trade off pure asynchronous semantics (where it makes sense) to circumvent the async state machine overhead
When implementing asynchronous support, there are methods that I changed to `async` such that even [bonehead exceptions](https://ericlippert.com/2008/09/10/vexing-exceptions/) would be caught by the state machine for async methods and placed on the returned task.
Those category of methods take the following form:
```csharp
async Task MethodAsync()
{
    VerifyYadaYada(); // Throws exception if expected conditions are not met
    await MethodAnotherAsync();
}
```
The `VerifyYadaYada` methods verifies for example that the stream is not disposed, or that a delta link is being written within a delta resource set, etc.
Without the `async` keyword, the exception would be raised directly rather than get placed on the returned task.
In this PR, I have dropped the `async` modifier from such methods for the following reasons:
- Unless there's a bug in our code, bonehead exceptions resulting from verification failures _should_ never happen.
- On the strength of the above point, we trade off pure asynchronous semantics - _exception being placed on the task returned from the method_ - for performance. There's little sense in incurring a performance cost (construction of state machine, instantiation of `Task` objects, etc.) for a condition that will almost never occur. If it occurred, the exception will still be caught by a method with the `async` modifier higher up in the call stack.

#### 2. Use of local functions to circumvent the async state machine overhead
Another category of methods where I dropped the `async` modifier are those where the asynchronous logic is run conditionally. By placing the asynchronous logic in a local function, it becomes possible to elide `async` and `await`. We get a performance benefit especially where the asynchronous logic is almost never executed. Such methods take the following form:
```csharp
async Task MethodAsync()
{
    if (true) // Check condition
    {
        // Asynchronous methods
    }
}
```
To avoid paying a higher cost when the condition is false, the rewritten method looks as follows:
```csharp
Task MethodAsync()
{
    if (true) // Check condition
    {
        return MethodInnerAsync();

        async Task MethodInnerAsync()
        {
            // Asynchronous methods
        }
    }

    return TaskUtils.CompletedTask; // Static property that returns a completed task instance
}
```
#### 3. Elide async/await where different branches of logic in a method makes it possible
This PR also elides `async` and `await` where each branch of logic in the asynchronous method consist of a single asynchronous method call. For example;
```csharp
async Task MethodAsync()
{
    if (true) // Condition check
    {
        await MethodAnotherAsync();
    }
    else if (true) // Another condition check
    {
        await MethodYetAnotherAsync();
    }

    throw new ODataException("Houston, we have a problem");
}
```
We rewrite these category of methods as follows:
```csharp
Task MethodAsync()
{
    if (true) // Condition check
    {
        return MethodAnotherAsync();
    }
    else if (true) // Another condition check
    {
        return MethodYetAnotherAsync();
    }

    return TaskUtils.GetFaultedTask(new ODataException("Houston, we have a problem"));
}
```

#### 4. Dropped use of expensive methods defined in `TaskUtils`
This PR also drops use of expensive methods defined in `TaskUtils` class - `FollowOnSuccessWithTask`, `FollowOnSuccessWith`, etc. Local functions that achieve the same purpose are introduced.

#### 5. Performance and memory usage improvements to particular methods defined in `TaskUtils`
There are asynchronous methods that we support but the logic in such methods do not currently perform any asynchronous operations.
For example, `CreateODataResourceWriterAsync` currently only initializes the writer to use for writing OData resources (same logic as `CreateODataResourceWriter`). It makes use of a method [`GetTaskForSynchronousOperation`](https://github.com/OData/odata.net/blob/7dcad74478debcfe54e93c63f47b7cb57f2d2e67/src/Microsoft.OData.Core/TaskUtils.cs#L108) defined in `TaskUtils` class. That method is used as follows:
```csharp
return TaskUtils.GetTaskForSynchronousOperation(
    () => this.CreateODataResourceWriterImplementation(navigationSource, resourceType));
```
Where `CreateODataResourceWriterImplementation` is a synchronous method.

I considered replacing all instances of `GetTaskForAsynchronousOperation` with `Task.FromResult` but held back since the `GetTaskForAsynchronousOperation` method does some exception handling and returns a faulted task for all exceptions except `OutOfMemoryException`.
The `GetTaskForAsynchronousOperation` method makes use of 2 helper methods (`GetCompletedTask` and `GetFaultedTask`).
I did some benchmarking to see whether using `Task.FromResult` and `Task.FromException` in the 2 helper methods would result into better performance and memory usage metrics.
```csharp
[MemoryDiagnoser]
public class CompletedTaskBenchmark
{
    private static readonly Type OutOfMemoryExceptionType = typeof(OutOfMemoryException);

    [Benchmark]
    public Task CompletedTaskUsingFromResultAsync()
    {
        return Task.FromResult("Task completed");
    }

    [Benchmark]
    public Task CompletedTaskUsingTaskCompletionSourceAsync()
    {
        return GetCompletedTask("Task completed");
    }

    [Benchmark]
    public Task CompletedTaskUsingGetTaskForSynchronousOperationAsync()
    {
        return GetTaskForSynchronousOperation(() => "Task completed");
    }

    [Benchmark]
    public Task CompletedTaskUsingGetTaskForSynchronousOperationRevisedAsync()
    {
        return GetTaskForSynchronousOperationRevised(() => "Task completed");
    }

    static Task<T> GetCompletedTask<T>(T value)
    {
        TaskCompletionSource<T> taskCompletionSource = new TaskCompletionSource<T>();
        taskCompletionSource.SetResult(value);
        return taskCompletionSource.Task;
    }

    internal static Task<T> GetFaultedTask<T>(Exception exception)
    {
        TaskCompletionSource<T> taskCompletionSource = new TaskCompletionSource<T>();
        taskCompletionSource.SetException(exception);
        return taskCompletionSource.Task;
    }

    static Task<T> GetTaskForSynchronousOperation<T>(Func<T> synchronousOperation)
    {
        try
        {
            T result = synchronousOperation();
            return GetCompletedTask(result);
        }
        catch (Exception ex)
        {
            Type type = ex.GetType();

            if (type == OutOfMemoryExceptionType)
            {
                throw;
            }

            return GetFaultedTask<T>(ex);
        }
    }

    static Task<T> GetTaskForSynchronousOperationRevised<T>(Func<T> synchronousOperation)
    {
        try
        {
            T result = synchronousOperation();
            return Task.FromResult(result);
        }
        catch (Exception ex)
        {
            Type type = ex.GetType();

            if (type == OutOfMemoryExceptionType)
            {
                throw;
            }

            return Task.FromException<T>(ex);
        }
    }
}
```
Below were the results:
```
|                                                       Method |      Mean |     Error |    StdDev |  Gen 0 | Allocated |
|------------------------------------------------------------- |----------:|----------:|----------:|-------:|----------:|
|                            CompletedTaskUsingFromResultAsync |  8.071 ns | 0.1713 ns | 0.2344 ns | 0.0172 |      72 B |
|                   CompletedTaskUsingTaskCompletedSourceAsync | 37.606 ns | 0.6161 ns | 0.5763 ns | 0.0229 |      96 B |
|        CompletedTaskUsingGetTaskForSynchronousOperationAsync | 40.309 ns | 0.4737 ns | 0.4431 ns | 0.0229 |      96 B |
| CompletedTaskUsingGetTaskForSynchronousOperationRevisedAsync | 13.357 ns | 0.2933 ns | 0.3709 ns | 0.0172 |      72 B |
```

From the above results, it's clear that the `GetCompletedTask` and `GetFaultedTask` as currently implemented are expensive. Using `TaskCompletionSource` to create a task instance has a higher cost than calling `Task.FromResult`.
I have refactored the `GetTaskForSynchronousOperation` to make use of both `Task.FromResult` and `Task.FromException` (where possible - `Task.FromException` is not supported across all frameworks that OData supports)

#### 6. Introduction of overloads of particular methods that accept delegates as parameters to prevent capturing of state from the enclosing context
The following 4 methods accept delegates as a parameter. In a lot of cases where they are used, they capture state from the enclosing context and as a consequence the C# compiler automatically generates a private "display class" that results into inadvertent allocations.
```csharp
Task WriteTopLevelPayloadAsync(Func<Task>) { /*...*/ } // ODataJsonLightSerializer

ODataContextUrlInfo WriteContextUriPropertyAsync(
    ODataPayloadKind,
    Func<ODataContextUrlInfo>,
    ODataContextUrlInfo,
    string) { /*...*/ } // ODataJsonLightSerializer

Task GetTaskForSynchronousOperation(Action) { /*...*/ } // TaskUtils

Task<T> GetTaskForSynchronousOperation(Func<T>) { /*...*/ } // TaskUtils
```

Below are examples of scenarios where "display classes" are created:
![image](https://user-images.githubusercontent.com/28291332/191093040-a12f6174-7ea0-4777-94a1-87e3ee7923d1.png)

![image](https://user-images.githubusercontent.com/28291332/191092805-c222871b-2983-46d5-a4ab-1d0cc141072f.png)

Since the above methods are quite heavily used, I introduced `protected` overloads that help prevent capturing state from the enclosing context. For `WriteTopLevelPayloadAsync` method for example, I introduced the following overloads:

```csharp
Task WriteTopLevelPayloadAsync<TArg>(Func<TArg, Task>, TArg) { /*...*/ }

Task WriteTopLevelPayloadAsync<TArg1, TArg2>(Func<TArg1, TArg2>, TArg1, TArg2) { /*...*/ }

Task WriteTopLevelPayloadAsync<TArg1, TArg2, TArg3>(Func<TArg1, TArg2, TArg3>, TArg1, TArg2, TArg3) { /*...*/ }
```
By making use of the overloads, the "display classes" are eliminated together with the associated allocation overheads
![image](https://user-images.githubusercontent.com/28291332/191092048-dc62ff86-1e90-46df-ba12-bc0e78f74eed.png)

No self-allocations in `ODataJsonLightResourceSerializer.WriteResourceSetContextUriAsync`
![image](https://user-images.githubusercontent.com/28291332/191094127-cfa50bb6-1c94-4d45-a930-cd79cfeef8d7.png)

#### Objects Allocations Report: Base vs Diff
```
summary:
better: 10, geomean: 1.199
worse: 11, geomean: 1.196
new (results in the diff that are not in the base): 12
missing (results in the base that are not in the diff): 32
total diff: 65

| Worse                                                    |          diff/base | Base Allocations | Diff Allocations | Modality|
| -------------------------------------------------------- | ------------------:| ----------------:| ----------------:| --------:|
| System.Func<,,,,>                                        |                  3 |             1.00 |             3.00 |         |
| System.Threading.ThreadPoolWorkQueueThreadLocals         |               1.25 |             4.00 |             5.00 |         |
| System.Threading.ThreadPoolWorkQueue.WorkStealingQueue   |               1.25 |             4.00 |             5.00 |         |
| System.Threading.Thread                                  |                1.2 |             5.00 |             6.00 |         |
| System.Threading.ThreadPoolWorkQueue.WorkStealingQueue[] |                1.2 |             5.00 |             6.00 |         |
| System.Action<,>                                         | 1.0263157894736843 |           114.00 |           117.00 |         |
| System.Threading.Tasks.Task<>                            | 1.0180722891566265 |           166.00 |           169.00 |         |
| System.Text.StringBuilder                                | 1.0081967213114753 |           244.00 |           246.00 |         |
| System.Func<,,>                                          | 1.0036231884057971 |          1656.00 |          1662.00 |         |
| System.Object[]                                          | 1.0033444816053512 |           598.00 |           600.00 |         |
| System.Char[]                                            |       1.0029296875 |          1024.00 |          1027.00 |         |

| Better                                                                          |          base/diff | Base Allocations | Diff Allocations | Modality|
| ------------------------------------------------------------------------------- | ------------------:| ----------------:| ----------------:| --------:|
| System.Threading.Tasks.TaskCompletionSource<>                                   |  3.176470588235294 |            54.00 |            17.00 |         |
| System.Action                                                                   | 1.1612903225806452 |           288.00 |           248.00 |         |
| System.Func<>                                                                   | 1.1336206896551724 |           263.00 |           232.00 |         |
| System.Runtime.CompilerServices.AsyncTaskMethodBuilder<>.AsyncStateMachineBox<> | 1.1290322580645162 |            35.00 |            31.00 |         |
| System.Threading.ContextCallback                                                | 1.1111111111111112 |            20.00 |            18.00 |         |
| System.Action<>                                                                 |  1.110655737704918 |           271.00 |           244.00 |         |
| System.Func<,>                                                                  | 1.0348178137651822 |          1278.00 |          1235.00 |         |
| System.Byte[]                                                                   | 1.0165289256198347 |           123.00 |           121.00 |         |
| System.SByte[]                                                                  | 1.0055137844611528 |          8024.00 |          7980.00 |         |
| System.String                                                                   |  1.000027430703186 |         72913.00 |         72911.00 |         |

| New                                                                      | diff/base | Base Allocations | Diff Allocations | Modality|
| ------------------------------------------------------------------------ | --------- | ----------------:| ----------------:| -------- |
| System.Func<,,,>                                                         | N/A       |                  |             6.00 | N/A     |
| System.Action<,,,,>                                                      | N/A       |                  |             2.00 | N/A     |
| Microsoft.OData.JsonLight.ODataJsonLightCollectionSerializer.<>c         | N/A       |                  |             1.00 | N/A     |
| Microsoft.OData.JsonLight.ODataJsonLightOutputContext.<>c                | N/A       |                  |             1.00 | N/A     |
| Microsoft.OData.JsonLight.ODataJsonLightPropertySerializer.<>c           | N/A       |                  |             1.00 | N/A     |
| Microsoft.OData.JsonLight.ODataJsonLightResourceSerializer.<>c           | N/A       |                  |             1.00 | N/A     |
| Microsoft.OData.JsonLight.ODataJsonLightServiceDocumentSerializer.<>c    | N/A       |                  |             1.00 | N/A     |
| Microsoft.OData.JsonLight.ODataJsonLightValueSerializer.<>c              | N/A       |                  |             1.00 | N/A     |
| Microsoft.OData.JsonLight.ODataJsonLightWriter.<>c                       | N/A       |                  |             1.00 | N/A     |
| Microsoft.OData.MultipartMixed.ODataMultipartMixedBatchOutputContext.<>c | N/A       |                  |             1.00 | N/A     |
| Microsoft.OData.MultipartMixed.ODataMultipartMixedBatchWriter.<>c        | N/A       |                  |             1.00 | N/A     |
| Microsoft.OData.RawValueWriter.<>c                                       | N/A       |                  |             1.00 | N/A     |

| Missing                                                                          | diff/base | Base Allocations | Diff Allocations | Modality|
| -------------------------------------------------------------------------------- | --------- | ----------------:| ----------------:| -------- |
| System.Threading.Tasks.ContinuationTaskFromTask                                  | N/A       |            48.00 |                  | N/A     |
| Microsoft.OData.JsonLight.ODataJsonLightValueSerializer.<>c__DisplayClass19_0    | N/A       |           102.00 |                  | N/A     |
| Microsoft.OData.JsonLight.ODataJsonLightSerializer.<>c__DisplayClass6_0          | N/A       |            62.00 |                  | N/A     |
| System.Threading.Tasks.Task.ContingentProperties                                 | N/A       |            24.00 |                  | N/A     |
| System.Threading.Tasks.StandardTaskContinuation                                  | N/A       |            48.00 |                  | N/A     |
| Microsoft.OData.JsonLight.ODataJsonLightWriter.<>c__DisplayClass82_0             | N/A       |            19.00 |                  | N/A     |
| Microsoft.OData.TaskUtils.<>c__DisplayClass27_0<>                                | N/A       |            24.00 |                  | N/A     |
| Microsoft.OData.ODataMessage.<>c__DisplayClass19_0                               | N/A       |            20.00 |                  | N/A     |
| Microsoft.OData.TaskUtils.<>c__DisplayClass13_0<,>                               | N/A       |            20.00 |                  | N/A     |
| Microsoft.OData.ODataMessageWriter.<>c__DisplayClass102_0<>                      | N/A       |            13.00 |                  | N/A     |
| Microsoft.OData.JsonLight.ODataJsonLightResourceSerializer.<>c__DisplayClass20_0 | N/A       |             6.00 |                  | N/A     |
| Microsoft.OData.JsonLight.ODataJsonLightOutputContext.<>c__DisplayClass30_0      | N/A       |             6.00 |                  | N/A     |
| Microsoft.OData.JsonLight.ODataJsonLightResourceSerializer.<>c__DisplayClass21_0 | N/A       |             6.00 |                  | N/A     |
| Microsoft.OData.ODataMessageWriter.<>c__DisplayClass101_0                        | N/A       |             7.00 |                  | N/A     |
| Microsoft.OData.ODataWriterCore.<>c__DisplayClass64_0                            | N/A       |             6.00 |                  | N/A     |
| Microsoft.OData.ODataBatchOperationMessage.<>c__DisplayClass10_0                 | N/A       |             4.00 |                  | N/A     |
| Microsoft.OData.ODataBatchWriter.<>c__DisplayClass71_0                           | N/A       |             4.00 |                  | N/A     |
| Microsoft.OData.RawValueWriter.<>c__DisplayClass18_0                             | N/A       |             3.00 |                  | N/A     |
| Microsoft.OData.JsonLight.ODataJsonLightWriter.<>c__DisplayClass83_0             | N/A       |             1.00 |                  | N/A     |
| Microsoft.OData.JsonLight.ODataJsonLightOutputContext.<>c__DisplayClass26_0      | N/A       |             1.00 |                  | N/A     |
…