@KenitoInc 2022-08-17
Can we get some benchmarks for this change