@habbes 2022-08-23
I ran your branch against some of the benchmarks in the repo. I found that it sped up the async writer scenarios (by up to 8%, awesome!) but it also increased memory allocations:

**Before**


|           Method |                                           WriterName |       Mean |     Error |    StdDev |      Gen 0 |     Gen 1 |  Allocated |
----------------- |----------------------------------------------------- |-----------:|----------:|----------:|-----------:|----------:|-----------:|
 WriteToFileAsync |                             ODataMessageWriter-Async | 900.471 ms | 5.0685 ms | 4.7411 ms | 66000.0000 | 1000.0000 | 406,503 KB |
 WriteToFileAsync |              ODataMessageWriter-Utf8JsonWriter-Async | 487.397 ms | 0.5287 ms | 0.4945 ms | 44000.0000 | 1000.0000 | 273,429 KB |

**After**

|           Method |                                           WriterName |       Mean |     Error |    StdDev |      Gen 0 |     Gen 1 |  Allocated |
----------------- |----------------------------------------------------- |-----------:|----------:|----------:|-----------:|----------:|-----------:|
 WriteToFileAsync |                             ODataMessageWriter-Async | 857.137 ms | 2.1981 ms | 1.9486 ms | 67000.0000 | 1000.0000 | 408,658 KB |
 WriteToFileAsync |              ODataMessageWriter-Utf8JsonWriter-Async | 448.118 ms | 0.4156 ms | 0.3887 ms | 47000.0000 | 1000.0000 | 292,088 KB |

I compared allocation reports from the VS profiler and found that display classes account for in the modified async methods are part of the cause for extra allocations:

**Before**

![image](https://user-images.githubusercontent.com/8460169/186077281-80e56a95-f864-479b-aebc-b8f94af6e106.png)


**After**

![image](https://user-images.githubusercontent.com/8460169/186077203-eb558008-b175-4bd4-a216-7434add99674.png)

Here's a full report from the `ResultsComparer` tool that shows the difference in allocation size for different methods between master and your branch. You can investigate the methods in the **New** and **Worse** tables for potential regressions.

summary:
better: 13, geomean: 1.678
worse: 25, geomean: 1.754
new (results in the diff that are not in the base): 15
missing (results in the base that are not in the diff): 14
total diff: 67

| Worse                                                                            | diff/base          | Base Self Size (bytes) | Diff Self Size (bytes) | Modality|
| -------------------------------------------------------------------------------- | ------------------ | ----------------------:| ----------------------:| --------:|
| Microsoft.OData.Json.JsonWriter.WriteNameAsync()                                 | Infinity           |                   0.00 |                 544.00 |         |
| Microsoft.OData.ODataWriterCore.WriteStartAsync(Microsoft.OData.ODataNestedResou | Infinity           |                   0.00 |               38400.00 |         |
| Microsoft.OData.Json.JsonWriterAsyncExtensions.WritePrimitiveValueAsync(Microsof | Infinity           |                   0.00 |                  80.00 |         |
| Microsoft.OData.ODataMessageWriter<T>.WriteToOutputAsync.__WriteToOutputInnerAsy | Infinity           |                   0.00 |                1056.00 |         |
| Microsoft.OData.JsonLight.ODataJsonLightWriter.WriteResourceSetDeltaLinkAsync(Sy | Infinity           |                   0.00 |                 128.00 |         |
| Microsoft.OData.JsonLight.ODataJsonLightWriter.WriteResourceSetNextLinkAsync(Sys | Infinity           |                   0.00 |               16160.00 |         |
| Microsoft.OData.Json.ODataJsonWriterUtils.StartJsonPaddingIfRequiredAsync(Micros | Infinity           |                   0.00 |                  64.00 |         |
| Microsoft.OData.JsonLightInstanceAnnotationWriter.WriteInstanceAnnotationsAsync( | Infinity           |                   0.00 |              115488.00 |         |
| Microsoft.OData.ODataWriterCore.WriteStartResourceImplementationAsync()          | 28.36842105263158  |                 152.00 |                4312.00 |         |
| Microsoft.OData.JsonLight.ODataJsonLightValueSerializer.WriteCollectionValueAsyn | 4.52               |                 600.00 |                2712.00 |         |
| Microsoft.OData.ODataResponseMessage.GetStreamAsync()                            | 3.375              |                 128.00 |                 432.00 |         |
| Microsoft.OData.ODataWriterCore.WriteStartAsync()                                | 2.573529411764706  |                 544.00 |                1400.00 |         |
| microsoft.aspnetcore.server.kestrel.core.il                                      | 2.5696681701030926 |               12416.00 |               31905.00 |         |
| Microsoft.OData.ODataMessageWriter<T>.MoveNext()                                 | 1.561307901907357  |                5872.00 |                9168.00 |         |
| Microsoft.OData.JsonLight.ODataJsonLightSerializer.WriteContextUriPropertyAsync( | 1.3289473684210527 |                 608.00 |                 808.00 |         |
| Microsoft.OData.ODataWriterCore.WriteStartResourceSetImplementationAsync()       | 1.2714570858283434 |                4008.00 |                5096.00 |         |
| microsoft.aspnetcore.http.abstractions.il                                        | 1.16               |                 200.00 |                 232.00 |         |
| Microsoft.OData.Json.JsonWriter.StartScopeAsync()                                | 1.1457203877115163 |               48696.00 |               55792.00 |         |
| Microsoft.OData.JsonLight.ODataJsonLightPropertySerializer.WritePropertyInfoAsyn | 1.0935251798561152 |                1112.00 |                1216.00 |         |
| Microsoft.OData.ODataWriterCore.WriteEndImplementationAsync.AnonymousMethod__199 | 1.0819672131147542 |                 488.00 |                 528.00 |         |
| Microsoft.OData.Json.JsonValueUtils.WriteEscapedJsonStringValueAsync()           | 1.078125           |                 512.00 |                 552.00 |         |
| microsoft.odata.core                                                             | 1.0406494651047598 |             3730824.00 |             3882480.00 |         |
| microsoft.aspnetcore.il                                                          | 1.025681700849486  |              958815.00 |              983439.00 |         |
| microsoft.aspnetcore.httpspolicy.il                                              | 1.0207206052057956 |               77990.00 |               79606.00 |         |
| microsoft.extensions.hosting.abstractions.il                                     | 1.0045951714422188 |              496173.00 |              498453.00 |         |

| Better                                                                           | base/diff          | Base Self Size (bytes) | Diff Self Size (bytes) | Modality|
| -------------------------------------------------------------------------------- | ------------------ | ----------------------:| ----------------------:| --------:|
| Microsoft.OData.ODataWriterCore.WriteStartAsync(Microsoft.OData.ODataResourceSet | Infinity           |                1056.00 |                   0.00 |         |
| Microsoft.OData.Json.JsonWriter.WriteNameAsync(string)                           | Infinity           |                1056.00 |                   0.00 |         |
| Microsoft.OData.JsonLight.ODataJsonLightPropertySerializer.WritePropertyAsync()  | 12.063829787234043 |                4536.00 |                 376.00 |         |
| Microsoft.OData.TaskUtils.GetTaskForSynchronousOperation<T>(System.Func<T>)      | 6.2                |                 248.00 |                  40.00 |         |
| microsoft.aspnetcore.mvc.il                                                      | 1.860798433894987  |               53230.00 |               28606.00 |         |
| Microsoft.OData.JsonLight.ODataJsonLightWriter.StartResourceSetAsync()           | 1.6914414414414414 |               12016.00 |                7104.00 |         |
| Microsoft.OData.ODataWriterCore.WriteStartResourceImplementationAsync.AnonymousM | 1.1646489104116222 |                3848.00 |                3304.00 |         |
| testserver                                                                       | 1.0358360906082973 |               40698.00 |               39290.00 |         |
| System.Private.CoreLib.il                                                        | 1.0145744493926645 |             6588335.00 |             6493693.00 |         |
| microsoft.extensions.logging.console.il                                          | 1.012705798138869  |               11318.00 |               11176.00 |         |
| system.private.uri.il                                                            | 1.012666244458518  |               12792.00 |               12632.00 |         |
| microsoft.aspnetcore.mvc.core.il                                                 | 1.0023767082590611 |              701792.00 |              700128.00 |         |
| experimentslib                                                                   | 1.0017169292402588 |              378066.00 |              377418.00 |         |

| New                                                                              | diff/base | Base Self Size (bytes) | Diff Self Size (bytes) | Modality|
| -------------------------------------------------------------------------------- | --------- | ----------------------:| ----------------------:| -------- |
| TestServer (PID: 24224)                                                          | N/A       |                        |               33810.00 | N/A     |
| Microsoft.OData.JsonLight.ODataJsonLightWriter.EndNestedResourceInfoWithContentA | N/A       |                        |               12800.00 | N/A     |
| Microsoft.OData.JsonLight.ODataJsonLightWriter.WriteResourceSetCountAsync(System | N/A       |                        |                9600.00 | N/A     |
| Microsoft.OData.ODataMessage.GetStreamAsync.__GetMessageStreamAsync0()           | N/A       |                        |                   0.00 | N/A     |
| Microsoft.OData.WriterValidator.ValidateTypeKind(Microsoft.OData.Edm.EdmTypeKind | N/A       |                        |                   0.00 | N/A     |
| Microsoft.OData.ODataWriterCore.CheckForNestedResourceInfoWithContentAsync.Anony | N/A       |                        |                   0.00 | N/A     |
| Microsoft.OData.UriParser.PropertySegment.ctor(Microsoft.OData.Edm.IEdmStructura | N/A       |                        |                   0.00 | N/A     |
| Microsoft.OData.JsonLightInstanceAnnotationWriter.ctor(Microsoft.OData.JsonLight | N/A       |                        |                   0.00 | N/A     |
| Microsoft.OData.JsonLight.ODataJsonLightWriterUtils.WriteInstanceAnnotationNameA | N/A       |                        |                   0.00 | N/A     |
| Microsoft.OData.Metadata.EdmLibraryExtensions.GetCollectionItemTypeName(string,  | N/A       |                        |                   0.00 | N/A     |
| Microsoft.OData.Json.JsonWriter.EndObjectScopeAsync()                            | N/A       |                        |                   0.00 | N/A     |
| Microsoft.OData.MediaTypeUtilsMatchInfoConcurrentCache.Add(MatchInfoCacheKey, Me | N/A       |                        |                   0.00 | N/A     |
| Microsoft.OData.ODataWriterCore.CheckForNestedResourceInfoWithContentAsync.Anony | N/A       |                        |                   0.00 | N/A     |
| microsoft.aspnetcore.diagnostics.il                                              | N/A       |                        |                 152.00 | N/A     |
| system.console.il                                                                | N/A       |                        |                  46.00 | N/A     |

| Missing                                                                          | diff/base | Base Self Size (bytes) | Diff Self Size (bytes) | Modality|
| -------------------------------------------------------------------------------- | --------- | ----------------------:| ----------------------:| -------- |
| TestServer (PID: 6408)                                                           | N/A       |               33794.00 |                        | N/A     |
| Microsoft.OData.ODataMessage.GetStreamAsync(System.Func<System.Threading.Tasks.T | N/A       |                5670.00 |                        | N/A     |
| Microsoft.OData.Json.JsonWriterAsyncExtensions.WritePrimitiveValueAsync()        | N/A       |                2192.00 |                        | N/A     |
| Microsoft.OData.ODataWriterCore.VerifyCanWriteStartResourceSetAsync()            | N/A       |                1400.00 |                        | N/A     |
| Microsoft.OData.TaskUtils.FollowOnSuccessWithContinuation<T>(System.Threading.Ta | N/A       |                1384.00 |                        | N/A     |
| Microsoft.OData.TaskUtils.FollowOnSuccessWithImplementation<T>(System.Threading. | N/A       |                1216.00 |                        | N/A     |
| Microsoft.OData.ODataMessage.GetStreamAsync.AnonymousMethod__0(System.Threading. | N/A       |                 666.00 |                        | N/A     |
| Microsoft.OData.TaskUtils.IgnoreExceptions(System.Threading.Tasks.Task)          | N/A       |                 208.00 |                        | N/A     |
| Microsoft.OData.TaskUtils.GetCompletedTask<T>(T)                                 | N/A       |                  48.00 |                        | N/A     |
| Microsoft.OData.TaskUtils.cctor()                                                | N/A       |                  24.00 |                        | N/A     |
| Microsoft.OData.ODataWriterCore.WriteStartAsync(Microsoft.OData.ODataResource)   | N/A       |                   0.00 |                        | N/A     |
| Microsoft.OData.ODataWriterCore.VerifyCanWriteStartResourceSetAsync(bool, Micros | N/A       |                   0.00 |                        | N/A     |
| microsoft.extensions.logging.abstractions.il                                     | N/A       |                  24.00 |                        | N/A     |
| system.net.security.il                                                           | N/A       |                  24.00 |                        | N/A     |