@habbes 2022-09-19
I've re-run the "SerializationComparisons" benchmarks and now I see improvements in both performance and memory of up to 10%. Great work.


**Before**

|           Method |                                           WriterName |       Mean |     Error |    StdDev |      Gen 0 |     Gen 1 |  Allocated |
|----------------- |----------------------------------------------------- |-----------:|----------:|----------:|-----------:|----------:|-----------:|
| WriteToFileAsync |                             ODataMessageWriter-Async | 911.242 ms | 8.6172 ms | 7.6389 ms | 66000.0000 | 1000.0000 | 403,823 KB |
| WriteToFileAsync |              ODataMessageWriter-Utf8JsonWriter-Async | 471.671 ms | 0.6272 ms | 0.5867 ms | 44000.0000 | 1000.0000 | 271,267 KB |

**After**

|           Method |                                           WriterName |       Mean |     Error |    StdDev |      Gen 0 |     Gen 1 |  Allocated |
|----------------- |----------------------------------------------------- |-----------:|----------:|----------:|-----------:|----------:|-----------:|
| WriteToFileAsync |                             ODataMessageWriter-Async | 816.809 ms | 2.3509 ms | 2.0840 ms | 59000.0000 | 1000.0000 | 362,984 KB |
| WriteToFileAsync |              ODataMessageWriter-Utf8JsonWriter-Async | 429.522 ms | 0.7683 ms | 0.7187 ms | 40000.0000 | 1000.0000 | 246,378 KB |